requirements.txt
PyMuPDF~=1.25.5
openai~=1.78.1
numpy~=2.2.5
scipy~=1.15.3
fitz~=0.0.1.dev2
python-dotenv~=1.1.0